requirements.txt
autopep8==1.4.3
Django==2.1.7
graphene==2.1.3
graphene-django==2.2.0
django-graphql-jwt==0.2.1
django-graphql-social-auth==0.1.4
social-auth-app-django==3.1.0
social-auth-core==3.1.0